Move kotlin util classes
