+++ subprojects/xsts/xsts-analysis/src/main/kotlin/hu/bme/mit/theta/xsts/analysis/util/XstsControlInitFunc.kt
@@ -14,10 +14,11 @@
  *  limitations under the License.
  */
 
-package hu.bme.mit.theta.xsts.analysis
+package hu.bme.mit.theta.xsts.analysis.util
 
 import hu.bme.mit.theta.analysis.Prec
 import hu.bme.mit.theta.analysis.unit.UnitState
+import hu.bme.mit.theta.xsts.analysis.XstsInitFunc
 
 fun <P : Prec> xstsControlInitFunc(): XstsInitFunc<UnitState, P> {
     return XstsInitFunc.create { _: P -> listOf<UnitState>(UnitState.getInstance()) }
